Moves Label component to another folder higher up in the directory structure
