+++ src/components/Label/__test__/Label.test.js
