+++ src/components/Label/index.js
