+++ src/signals/incident-management/components/DatePickerInput/index.js
@@ -4,7 +4,7 @@ import DatePicker from 'react-datepicker';
 import moment from 'moment';
 import 'react-datepicker/dist/react-datepicker.css';
 
-import Label from '../Label';
+import Label from 'components/Label';
 
 import './style.scss';
